Add HamburgerMenu component to about page and main page
